+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- PostgreSQL 11 Support
+- Support setting database flags (new configuration parameter: `database_flags`).
+- Support setting the storage auto resize limit (new configuration parameter: `auto_resize_limit`).
+
+### Fixed
+- CloudSQL instances can be configured with high availability again (new configuration parameter: `availability_type`).
+
+### Removed
+- Support for MySQL 5.5.
+- Support for first generation CloudSQL instances and their associated fields.
+
 ## [4.3.0] - 2019-08-26
 
 ### Fixed